Move canonicalization code around
